24_07_01 10:30 음성점수 및 총점수 평균가져오기 음성점수 및 총점수 평균 가져오지

+++ src/main/java/io/paioneer/nain/interview/jpa/entity/InterviewEntity.java
@@ -47,7 +47,7 @@ public class InterviewEntity {
     private double videoScore;
 
     @Column(name="VOICE_SCORE", nullable = false)
-    private int voiceScore;
+    private double voiceScore;
 
     @Column(name="ITV_DATE", nullable = false)
     private Date itvDate;
